% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_density_utils.R
\name{construct_density_matrix}
\alias{construct_density_matrix}
\title{Construct and evaluate a simplex density as a grid-based color map}
\usage{
construct_density_matrix(
  density_function,
  args,
  resolution = 101,
  col_scale = "linear"
)
}
\arguments{
\item{density_function}{Density function that is evaluated at each grid point}

\item{args}{arguments to the density function}

\item{resolution}{number of grid points along axes}

\item{col_scale}{scale mapping, e.g. "linear" or "sqrt"}
}
\value{
A matrix of color codes
}
\description{
Construct and evaluate a simplex density as a grid-based color map
}
