% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_grid_utils.R
\name{construct_simplex_grid}
\alias{construct_simplex_grid}
\title{Construct a simplex grid of specified resolution}
\usage{
construct_simplex_grid(n_x, n_y)
}
\arguments{
\item{n_x}{number of grid points along x-axis}

\item{n_y}{number of grid points along y-axis}
}
\value{
Grid of points within the 2-simplex
}
\description{
Construct a simplex grid of specified resolution
}
