% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dot}
\alias{dot}
\title{Calculate the dot product of two vectors.}
\usage{
dot(u, v)
}
\arguments{
\item{u}{A numeric vector}

\item{v}{A numeric vector}
}
\value{
The dot product (scalar) of \code{u} and \code{v}
}
\description{
Calculate the dot product of two vectors.
}
\examples{
u = c(1, 10, 100)
v = c(-1, 0, 1)
dot(u, v)
}
