% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_simplex.R
\name{geom_simplex}
\alias{geom_simplex}
\title{Draw points on a simplex from barycentric coordinates}
\usage{
geom_simplex(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  stat = "identity",
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{stat}{}

\item{inherit.aes}{}

\item{...}{}
}
\value{

}
\description{
Draw points on a simplex from barycentric coordinates
}
