% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_simplex_canvas.R
\name{geom_simplex_canvas}
\alias{geom_simplex_canvas}
\title{Draw Simplex Canvas}
\usage{
geom_simplex_canvas(
  mapping = NULL,
  data = data.frame(0),
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  ...
)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{stat}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}

\item{...}{}
}
\value{

}
\description{
Draw Simplex Canvas
}
