% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_simplex_heatmap.R
\name{geom_simplex_heatmap}
\alias{geom_simplex_heatmap}
\title{Draw a simplex heatmap from a matrix of color values}
\usage{
geom_simplex_heatmap(
  mapping = NULL,
  data = 0,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  ...
)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{stat}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}

\item{...}{}
}
\value{

}
\description{
Draw a simplex heatmap from a matrix of color values
}
