% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_list_column}
\alias{make_list_column}
\title{Helper function to bind columns into a single list column}
\usage{
make_list_column(...)
}
\arguments{
\item{...}{columns that shall be turned into the list column}
}
\value{
single list column which consits of the specified columns
}
\description{
Helper function to bind columns into a single list column
}
