% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_density_utils.R
\name{construct_density_matrix}
\alias{construct_density_matrix}
\title{Construct and evaluate a simplex density as a grid-based color map}
\usage{
construct_density_matrix(density_function, args)
}
\arguments{
\item{density_function}{}

\item{args}{}
}
\value{
A matrix of color codes
}
\description{
Construct and evaluate a simplex density as a grid-based color map
}
