% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_grid_utils.R
\name{construct_simplex_grid}
\alias{construct_simplex_grid}
\title{Construct a simplex grid of specified resolution}
\usage{
construct_simplex_grid(n_x, n_y)
}
\arguments{
\item{n_x}{}

\item{n_y}{}
}
\value{

}
\description{
Construct a simplex grid of specified resolution
}
