% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_list_column}
\alias{make_list_column}
\title{Helper function to bind columns into a single list column}
\usage{
make_list_column(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Helper function to bind columns into a single list column
}
