% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_simplex_density.R
\name{stat_simplex_density}
\alias{stat_simplex_density}
\title{Draw a simplex density heatmap from distributional parameters}
\usage{
stat_simplex_density(
  mapping = NULL,
  data = NULL,
  geom = "simplex_heatmap",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  fun = NULL,
  args = NULL,
  ...
)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{geom}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}

\item{fun}{}

\item{args}{}

\item{...}{}
}
\value{

}
\description{
Draw a simplex density heatmap from distributional parameters
}
